Fix: Mockのみに存在するint変換を削除

+++ voicevox_engine/dev/synthesis_engine/mock.py
@@ -119,8 +119,5 @@ class MockSynthesisEngine(SynthesisEngineBase):
         logger = getLogger("uvicorn")  # FastAPI / Uvicorn 内からの利用のため
         logger.info("[Mock] input text: %s" % text)
         wave, sr = tts(text)
-        wave = resample(
-            wave.astype("int16"),
-            24000 * len(wave) // 48000,
-        )
-        return wave.astype("int16")
+        wave = resample(wave, 24000 * len(wave) // 48000)
+        return wave
